pyproject.toml
[tool.poetry]
name = "chatroom"
version = "0.1.0"
description = "Real-time chatroom backend with FastAPI, Socket.IO, and in-memory SQLite"
authors = ["Shweta Seth <shwetadseth@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
python-socketio = { extras = ["asgi"], version = "^5.11.0" }
sqlalchemy = "^2.0.30"
aiohttp = "^3.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.3"
httpx = "^0.27.0"
python-socketio = { version = "^5.11.0", extras = ["client"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
